media-sound/pulseaudio: Fixed rm error.
Closes: https://bugs.gentoo.org/658704
Package-Manager: Portage-2.3.40, Repoman-2.3.9

+++ media-sound/pulseaudio/pulseaudio-12.0-r1.ebuild
@@ -282,7 +282,9 @@ multilib_src_install_all() {
 		systemd_newtmpfilesd "${FILESDIR}/${PN}.tmpfiles" "${PN}.conf"
 	else
 		# Prevent warnings when system-wide is not used, bug #447694
-		rm "${ED%/}"/etc/dbus-1/system.d/pulseaudio-system.conf || die
+		if use dbus ; then
+			rm "${ED%/}"/etc/dbus-1/system.d/pulseaudio-system.conf || die
+		fi
 	fi
 
 	if use zeroconf ; then
